cli service refactoring
