+++ cmd/service/detail.go
@@ -8,15 +8,15 @@ import (
 
 // Detail returns all the details of a service
 var Detail = &cobra.Command{
-	Use:               "detail SERVICE_ID",
+	Use:               "detail SERVICE_FOLDER",
 	Short:             "Show details of a published service",
 	Args:              cobra.MinimumNArgs(1),
-	Example:           "mesg-cli service detail SERVICE_ID",
+	Example:           "mesg-cli service detail SERVICE_FOLDER",
 	Run:               detailHandler,
 	DisableAutoGenTag: true,
 }
 
 func detailHandler(cmd *cobra.Command, args []string) {
-	// TODO Fetch details and display
-	fmt.Println("service details : ", args)
+	service := loadService(defaultPath(args))
+	fmt.Println("service details : ", service)
 }
