+++ cmd/service/utils.go
@@ -5,6 +5,7 @@ import (
 	"os"
 
 	"github.com/logrusorgru/aurora"
+	"github.com/mesg-foundation/core/service"
 )
 
 // Set the default path if needed
@@ -20,4 +21,23 @@ func handleError(err error) {
 		fmt.Println(aurora.Red(err))
 		os.Exit(0)
 	}
+}
+
+func loadService(path string) (importedService *service.Service) {
+	importedService, err := service.ImportFromPath(path)
+	if err != nil {
+		fmt.Println(aurora.Red(err))
+		fmt.Println("Run the command 'service validate' to get detailed errors")
+		os.Exit(0)
+	}
+	return
+}
+
+func startService(service *service.Service) {
+	_, err := service.Start()
+	if err != nil {
+		fmt.Println(aurora.Red(err))
+		os.Exit(0)
+	}
+	fmt.Println(aurora.Green("Service started"))
 }
